extract repository interfaces
